GET /api/stats/activity platform activity route
The new route allows admins to retrieve the number of:
* planned excursions
* started installations
* installation events
* created user accounts

The data is aggregated by hour, day, week or month.

A time interval can optionally be specified. It defaults to:
* 24 hours for the hour interval
* 30 days for the day interval
* 26 weeks (half a year) for the week interval
* 24 months for the month interval

That way the number of data points remains similar (24-30).

Empty hours/days/weeks/months are sent in the response with a count of
zero.